- split nvp_driver and optimize_nvp_driver classes
  NB: you need to edit your quantum.conf file like so:

  before:
  [QUARK]
  net_driver = quark.drivers.nvp_driver.OptimizedNVPDriver

  after:
  [QUARK]
  net_driver = quark.drivers.optimized_nvp_driver.OptimizedNVPDriver

  note the new optimized_nvp_driver module instead of nvp_driver

- add swp files to gitignore
