+++ .gitignore
@@ -37,3 +37,5 @@ nosetests.xml
 
 # virtualenv
 .venv
+
+*.swp
