+++ quark/tests/test_optimized_nvp_driver.py
@@ -16,16 +16,16 @@
 import contextlib
 import mock
 
-import quark.drivers.nvp_driver
+import quark.drivers.optimized_nvp_driver
 import quark.tests.test_nvp_driver as test_nvp_driver
 
 
 class TestOptimizedNVPDriver(test_nvp_driver.TestNVPDriver):
     def setUp(self):
         if not hasattr(self, 'driver'):
-            self.driver = quark.drivers.nvp_driver.OptimizedNVPDriver()
+            self.driver = quark.drivers.optimized_nvp_driver.OptimizedNVPDriver()
         super(TestOptimizedNVPDriver, self).setUp()
-        self.d_pkg = "quark.drivers.nvp_driver.OptimizedNVPDriver"
+        self.d_pkg = "quark.drivers.optimized_nvp_driver.OptimizedNVPDriver"
         self.context.session.add = mock.Mock(return_value=None)
         self.net_id = "12345678-1234-0000-1234-123412341234"
